Refactor: reorganize AST parsing logic, remove unused files, and clean up tokenizer
